RISC-V: fix bug introduced in `d6f6cc1ada1`
This bug caused factorial test to fail. It was caused by bad
refactoring in `::BBEndEvaluator`:

    if (NULL == block->getNextBlock())
       {
       TR::Instruction *lastInstruction = cg->getAppendInstruction();
-      if (lastInstruction->getOpCodeValue() == TR::InstOpCode::bl
+      if (lastInstruction->getOpCodeValue() == TR::InstOpCode::_jal
               && lastInstruction->getNode()->getSymbolReference()->getReferenceNumber() == TR_aThrow)
          {
          lastInstruction = generateInstruction(cg, TR::InstOpCode::bad, node, lastInstruction);

PowerPC code is very similar. I don't know the reason for this code, however on Aarch64 and
PowerPC for normal gotos only `b` is used. There's no such thing in RISC-V, `jal` and `jalr`
always "link", when the return address is not needed (such an in goto), it links to `x0` / `zero`.
Therefore we should probably also test `rd` of the instruction being `x0` / `zero`.

This will be a lot easier once instruction herarchy is revamped so this commit just
adds a hack and will be revisited later after instruction refactoring.